EditDialog: refactor - extract utils
