+++ src/components/FeaturePanel/EditDialog/EditContent/FeatureEditSection/AddMemberForm.tsx
@@ -1,5 +1,4 @@
 import { useState } from 'react';
-import { getPresetKey } from '../../context/useEditItems';
 import { getApiId } from '../../../../../services/helpers';
 import { getOsmElement } from '../../../../../services/osm/quickFetchFeature';
 import { useCurrentItem, useEditContext } from '../../context/EditContext';
@@ -12,6 +11,7 @@ import { useMapStateContext, View } from '../../../../utils/MapStateContext';
 import { getPresetTranslation } from '../../../../../services/tagging/translations';
 import { fetchFreshItem, getNewNodeItem } from '../../context/itemsHelpers';
 import { DataItem, EditDataItem, Members } from '../../context/types';
+import { getPresetKey } from '../../context/utils';
 
 const getLastNode = (members: Members) => {
   const lastNode = members
