+++ src/components/FeaturePanel/EditDialog/EditContent/useHandleItemClick.tsx
@@ -1,4 +1,3 @@
-import { isInItems } from '../context/useEditItems';
 import React from 'react';
 import { useCurrentItem, useEditContext } from '../context/EditContext';
 import { getApiId, getShortId } from '../../../../services/helpers';
@@ -6,6 +5,8 @@ import { fetchFreshItem } from '../context/itemsHelpers';
 import { Setter } from '../../../../types';
 import { Feature } from '../../../../services/types';
 import { DataItem, EditDataItem } from '../context/types';
+
+import { isInItems } from '../context/utils';
 
 export const useHandleItemClick = (setIsExpanded: Setter<boolean>) => {
   const { addItem, items, setCurrent } = useEditContext();
