+++ src/components/FeaturePanel/EditDialog/context/useEditItems.tsx
@@ -2,16 +2,10 @@ import { LonLat } from '../../../../services/types';
 import { getApiId } from '../../../../services/helpers';
 import { Setter } from '../../../../types';
 import { useCallback, useMemo, useState } from 'react';
-import { not, publishDbgObject } from '../../../../utils';
-import { findPreset } from '../../../../services/tagging/presets';
+import { publishDbgObject } from '../../../../utils';
 import { getPresetTranslation } from '../../../../services/tagging/translations';
-import { getNewId } from '../../../../services/getCoordsFeature';
-import { fetchParentFeatures } from '../../../../services/osm/fetchParentFeatures';
-import { fetchWays } from '../../../../services/osm/fetchWays';
-import { addEmptyOriginalState, fetchFreshItem } from './itemsHelpers';
 import { isEqual } from 'lodash';
 import {
-  ConvertToRelation,
   DataItem,
   EditDataItem,
   Members,
@@ -20,25 +14,8 @@ import {
   SetTagsEntries,
   TagsEntries,
 } from './types';
-
-export const isInItems = (items: DataItem[], shortId: string) =>
-  items.some((item) => item.shortId === shortId);
-
-const findInItems = (items: DataItem[], shortId: string) =>
-  items.find((item) => item.shortId === shortId);
-
-export const getPresetKey = ({ shortId, tagsEntries }: DataItem) => {
-  const tags = Object.fromEntries(tagsEntries);
-  const osmId = getApiId(shortId);
-  const preset = findPreset(osmId.type, tags);
-  return preset.presetKey;
-};
-
-const getName = (d: DataItem): string | undefined =>
-  d.tagsEntries.find(([k]) => k === 'name')?.[1];
-
-const getLabel = (newRelation: DataItem) =>
-  getName(newRelation) ?? newRelation.shortId;
+import { convertToRelationFactory } from './convertToRelationFactory';
+import { getName, getPresetKey } from './utils';
 
 const someNameHasChanged = (prevData: DataItem[], newData: DataItem[]) => {
   const prevNames = prevData.map((d) => getName(d));
@@ -86,124 +63,6 @@ const setDataItemFactory =
       return newData;
     });
   };
-
-const updateMemberLinks = (
-  item: DataItem,
-  oldShortId: string,
-  newRelation: DataItem,
-) => {
-  if (item.shortId === newRelation.shortId) {
-    return item;
-  }
-
-  // update member id in every parent
-  return {
-    ...item,
-    members: item.members?.map((member) =>
-      member.shortId === oldShortId
-        ? {
-            ...member,
-            shortId: newRelation.shortId,
-            label: getLabel(newRelation),
-          }
-        : member,
-    ),
-  };
-};
-
-// TODO we may lose some custom tags, if converting new item (it is deleted afterwards)
-//  remove climbing=* and find if this still matches some other preset (only keep it in that scenario)  (?)
-const copiedClimbingTags = ([k, v]) =>
-  k.startsWith('name') ||
-  k.startsWith('climbing') ||
-  k.startsWith('description') ||
-  k.startsWith('wikimedia_commons') ||
-  k.startsWith('website') ||
-  (k === 'sport' && v === 'climbing');
-
-const toBeRemovedTags = ([k, v]) =>
-  k.startsWith('climbing') || (k === 'sport' && v === 'climbing');
-
-const isNew = (item: DataItem) => item.shortId.includes('-');
-
-const getConversionTags = (node: DataItem) => {
-  const tagsToCopy = node.tagsEntries.filter(copiedClimbingTags);
-  const restTags = node.tagsEntries.filter(not(copiedClimbingTags));
-
-  const keepNode = !isNew(node) && restTags.length > 0;
-  const keptTags = keepNode
-    ? node.tagsEntries.filter(not(toBeRemovedTags))
-    : [];
-
-  return { tagsToCopy, keepNode, keptTags };
-};
-
-const fetchParentItems = async (shortId: string) => {
-  const parentFeatures = await fetchParentFeatures(getApiId(shortId)); // without memberFeatures
-  return await Promise.all(
-    parentFeatures.map((feature) => fetchFreshItem(feature.osmMeta)), // we need full item (with members)
-  );
-};
-
-const convertToRelationFactory = (
-  setData: Setter<DataItem[]>,
-  shortId: string,
-): ConvertToRelation => {
-  // should work only for node - new or existing
-
-  return async () => {
-    const [parentItems, waysFeatures] = await Promise.all([
-      fetchParentItems(shortId),
-      fetchWays(getApiId(shortId)),
-    ]);
-
-    if (waysFeatures.length > 0) {
-      throw new Error(`Can't convert node ${shortId} which is part of a way.`); // TODO duplicate the node ?
-    }
-
-    const newShortId = `r${getNewId()}`;
-    setData((prevData) => {
-      const node = findInItems(prevData, shortId);
-      const { tagsToCopy, keepNode, keptTags } = getConversionTags(node);
-
-      const newRelation: DataItem = addEmptyOriginalState({
-        shortId: newShortId,
-        version: undefined,
-        tagsEntries: Object.entries(
-          Object.fromEntries([
-            ['type', 'site'],
-            ['site', 'climbing'],
-            ...tagsToCopy,
-          ]),
-        ),
-        toBeDeleted: false,
-        nodeLonLat: node.nodeLonLat, // will be used later for first node TODO rename it to make it clear
-        nodes: undefined,
-        members: keepNode ? [{ shortId, role: '', label: getLabel(node) }] : [],
-      });
-
-      const newData = prevData.map((item) =>
-        item.shortId === shortId
-          ? keepNode
-            ? { ...item, tagsEntries: keptTags }
-            : { ...item, toBeDeleted: true, tagsEntries: [] }
-          : item,
-      );
-      newData.push(newRelation);
-
-      // add all parent relations which are not already there
-      newData.push(
-        ...parentItems.filter((parent) => !isInItems(newData, parent.shortId)),
-      );
-
-      return newData.map((item) =>
-        updateMemberLinks(item, shortId, newRelation),
-      );
-    });
-
-    return newShortId;
-  };
-};
 
 const setTagsEntriesFactory =
   (setDataItem: SetDataItem, tagsEntries: TagsEntries): SetTagsEntries =>
@@ -322,5 +181,3 @@ export const useEditItems = () => {
 
   return { items, addItem, removeItem };
 };
-
-export { convertToRelationFactory };
